新增實體課程促銷功能 (#88)
* 新增實體課程促銷功能

* 新增實體課程功能“

+++ web-admin/src/main/java/com/phyCouPromotionDetaiil/controller/phyCouPromotionDetailServlet.java
@@ -1,4 +1,4 @@
-package com.phyCouPromotion.controller;
+package com.phyCouPromotionDetaiil.controller;
 
 import java.io.*;
 import java.util.*;
@@ -17,8 +17,8 @@ import com.phyCouPromotionDetail.model.PhyCouPromotionDetailVO;
 
 
 
-@WebServlet("/phyCouPromotion/promotion")
-public class phyCouPromotionServlet extends HttpServlet {
+@WebServlet("/phyCouPromotionDetail/promotionDetail")
+public class phyCouPromotionDetailServlet extends HttpServlet {
 
 	private static final long serialVersionUID = 1L;
 
@@ -99,33 +99,32 @@ public class phyCouPromotionServlet extends HttpServlet {
 			
 				/***************************1.接收請求參數****************************************/
 				Integer project_no = Integer.valueOf(req.getParameter("project_no"));
+				Integer course_no = Integer.valueOf(req.getParameter("course_no"));
 				
 				/***************************2.開始查詢資料****************************************/
-				PhyCouPromotionService phyCouPromotionSvc = new PhyCouPromotionService();
-				PhyCouPromotionVO phyCouPromotionVO = phyCouPromotionSvc.getOnePro(project_no);
-				Set<PhyCouPromotionDetailVO> set = phyCouPromotionVO.getPhyCouPromotionDetails();
-				
-				PhyCouPromotionVO vo = new PhyCouPromotionVO();				
-				vo.setProject_no(phyCouPromotionVO.getProject_no());
-				vo.setProject_name(phyCouPromotionVO.getProject_name());
-				vo.setStart_date(phyCouPromotionVO.getStart_date());
-				vo.setEnd_date(phyCouPromotionVO.getEnd_date());
-				vo.setProm_description(phyCouPromotionVO.getProm_description());
-				vo.setProm_status(phyCouPromotionVO.getProm_status());
+				PhyCouPromotionDetailService phyCouPromotionDetailSvc = new PhyCouPromotionDetailService();
+				List<PhyCouPromotionDetailVO> list = phyCouPromotionDetailSvc.getOneDetail(project_no, course_no);
+//				Set<PhyCouPromotionDetailVO> set = phyCouPromotionVO.getPhyCouPromotionDetails();
+				Integer prom_price = null;
+				for (PhyCouPromotionDetailVO one : list ) {
+//				PhyCouPromotionDetailVO vo = new PhyCouPromotionDetailVO();				
+//				vo.setProject_no(one.getPhyCouPromotionVO().getProject_no());
+//				vo.setProject_name(phyCouPromotionVO.getProject_name());
+//				vo.setStart_date(phyCouPromotionVO.getStart_date());
+//				vo.setEnd_date(phyCouPromotionVO.getEnd_date());
+//				vo.setProm_description(phyCouPromotionVO.getProm_description());
+				prom_price = one.getProm_price();
+				}
 				
 				/***************************3.查詢完成,準備轉交(Send the Success view)************/
-				String param = "?project_no="    +phyCouPromotionVO.getProject_no()+
-						       "&project_name="  +phyCouPromotionVO.getProject_name()+
-						       "&start_date="    +phyCouPromotionVO.getStart_date()+
-						       "&end_date=" +phyCouPromotionVO.getEnd_date()+
-						       "&prom_description="+phyCouPromotionVO.getProm_description()+
-						       "&prom_status="  +phyCouPromotionVO.getProm_status()+
-						       "&update_time="  +phyCouPromotionVO.getUpdate_time();
+				String param = "?project_no="    +project_no+
+						       "&course_no="  +course_no+
+						       "&prom_price="  +prom_price;
 
 					       
-				req.setAttribute("phyCouPromotionVO", vo);        
-				req.setAttribute("set", set);        
-				String url = "/phyCouPromotion/update_pro_input.jsp"+param;
+//				req.setAttribute("phyCouPromotionDetailVO", vo);        
+//				req.setAttribute("set", set);        
+				String url = "/phyCouPromotion/update_proDetail_input.jsp"+param;
 				RequestDispatcher successView = req.getRequestDispatcher(url);
 				successView.forward(req, res);
 		}
@@ -140,77 +139,9 @@ public class phyCouPromotionServlet extends HttpServlet {
 		
 				/***************************1.接收請求參數 - 輸入格式的錯誤處理**********************/
 			Integer project_no = Integer.valueOf(req.getParameter("project_no").trim());		
-
-			String project_name = req.getParameter("project_name");
-			if (project_name == null || project_name.trim().length() == 0) {
-				errorMsgs.put("project_name", "專案名稱 請勿空白");
-				System.out.println("專案名稱 請勿空白");
-			}
-			
-			java.sql.Date start_date = null;
-			try {				
-				start_date = java.sql.Date.valueOf(req.getParameter("start_date").trim());
-			} catch (IllegalArgumentException e) {
-				start_date = new java.sql.Date(System.currentTimeMillis());
-				errorMsgs.put("start_date", "請輸入日期!");
-				System.out.println("請輸入日期!");
-			}
-			
-			java.sql.Date end_date = null;
-			try {				
-				end_date = java.sql.Date.valueOf(req.getParameter("end_date").trim());
-			} catch (IllegalArgumentException e) {
-				end_date = new java.sql.Date(System.currentTimeMillis());
-				errorMsgs.put("end_date", "請輸入日期!");
-				System.out.println("請輸入日期!");
-			}
-			
-			String prom_description = req.getParameter("prom_description");
-			if (prom_description == null || prom_description.trim().length() == 0) {
-				errorMsgs.put("prom_description", "促銷活動敘述 請勿空白");
-				System.out.println("促銷活動敘述 請勿空白");
-			}
-			
-			Integer prom_status = null;
-			try {		
-				prom_status = Integer.valueOf(req.getParameter("prom_status").trim()); 
-			} catch (Exception e) {
-				errorMsgs.put("prom_status", "專案狀態 請填數字");
-				System.out.println("專案狀態 請填數字");
-			}
-			
-			java.sql.Date update_time = null;
-			try {				
-				update_time = java.sql.Date.valueOf(req.getParameter("update_time").trim());
-			} catch (IllegalArgumentException e) {
-				update_time = new java.sql.Date(System.currentTimeMillis());
-				errorMsgs.put("update_time", "請輸入日期!");
-				System.out.println("請輸入日期!");
-			}
-			
-			String tep_proCous = null;
-			tep_proCous= req.getParameter("proCous");
-			System.out.println("======================================================");
-			System.out.println(tep_proCous);
-			System.out.println("======================================================");
-			if (tep_proCous == null || tep_proCous.trim().length() == 0) {
-				errorMsgs.put("proCous", "促銷課程編號 請勿空白");
-				System.out.println("促銷課程編號 請勿空白");
-			}
-			String[] proCous = tep_proCous.split(",");
-	
-			for ( int i = 0; i < proCous.length; i++)
-			{
-				try {
-					proCous[i] = proCous[i].trim();
-					Integer.valueOf(proCous[i]);
-				} catch (Exception e) {
-					errorMsgs.put("proCous", "促銷課程編號 請填數字");
-					System.out.println("促銷課程編號 請填數字");
-				}
-			}
-				
-			Integer prom_price = null;
+			Integer course_no = Integer.valueOf(req.getParameter("course_no").trim());		
+					
+		    Integer prom_price = null;
 			try {		
 				prom_price = Integer.valueOf(req.getParameter("prom_price").trim()); 
 			} catch (Exception e) {
@@ -223,36 +154,33 @@ public class phyCouPromotionServlet extends HttpServlet {
 //				req.setAttribute("phyCouVO", phyCouVO); 
 				System.out.println("error");
 				System.out.println("================================================");
-//				req.setAttribute("tep_proCous", tep_proCous);
-//				req.setAttribute("prom_price", prom_price);
-//				req.setAttribute("update_time", update_time);
+				req.setAttribute("prom_price", prom_price);
+
 				
 				RequestDispatcher failureView = req
-						.getRequestDispatcher("/phyCouPromotion/update_pro_input.jsp");
+						.getRequestDispatcher("/phyCouPromotionDetao;/update_proDetail_input.jsp");
 				failureView.forward(req, res);
 				return;
 			}
 				
 				/***************************2.開始修改資料*****************************************/
-				PhyCouPromotionService phyCouPromotionSvc = new PhyCouPromotionService();
-				PhyCouPromotionVO phyCouPromotionVO = phyCouPromotionSvc.updatePhyCouPromotion(project_no, project_name, start_date, end_date, prom_description, prom_status, proCous, prom_price);
-				System.out.println("==========================================");
-				System.out.println(phyCouPromotionVO);
-				System.out.println("==========================================");
-				Set<PhyCouPromotionDetailVO> set = phyCouPromotionVO.getPhyCouPromotionDetails();
-				
-				PhyCouPromotionVO vo = new PhyCouPromotionVO();				
-				vo.setProject_no(phyCouPromotionVO.getProject_no());
-				vo.setProject_name(phyCouPromotionVO.getProject_name());
-				vo.setStart_date(phyCouPromotionVO.getStart_date());
-				vo.setEnd_date(phyCouPromotionVO.getEnd_date());
-				vo.setProm_description(phyCouPromotionVO.getProm_description());
-				vo.setProm_status(phyCouPromotionVO.getProm_status());
-				vo.setUpdate_time(phyCouPromotionVO.getUpdate_time());
+				PhyCouPromotionDetailService phyCouPromotionDetailSvc = new PhyCouPromotionDetailService();
+				PhyCouPromotionDetailVO phyCouPromotionDetailVO = phyCouPromotionDetailSvc.updateProDetail(project_no, course_no, prom_price);
+								
+			    Integer prom_price1 = phyCouPromotionDetailVO.getProm_price();
+//				vo.setProject_no(phyCouPromotionVO.getProject_no());
+//				vo.setProject_name(phyCouPromotionVO.getProject_name());
+//				vo.setStart_date(phyCouPromotionVO.getStart_date());
+//				vo.setEnd_date(phyCouPromotionVO.getEnd_date());
+//				vo.setProm_description(phyCouPromotionVO.getProm_description());
+//				vo.setProm_status(phyCouPromotionVO.getProm_status());
+//				vo.setUpdate_time(phyCouPromotionVO.getUpdate_time());
 				/***************************3.修改完成,準備轉交(Send the Success view)*************/
-				req.setAttribute("phyCouPromotionVO", vo); 
-				req.setAttribute("set", set); 
-				String url = "/phyCouPromotion/listOnePro.jsp";
+				req.setAttribute("project_no", project_no); 
+				req.setAttribute("course_no", course_no); 
+				req.setAttribute("prom_price", prom_price1); 
+				
+				String url = "/phyCouPromotion/listOneProDetail.jsp";
 				RequestDispatcher successView = req.getRequestDispatcher(url); // 修改成功後,轉交listOneEmp.jsp
 				successView.forward(req, res);
 		}
@@ -341,7 +269,7 @@ public class phyCouPromotionServlet extends HttpServlet {
         
 		
 		
-		if ("delete".equals(action)) { // 來自listAllEmp.jsp
+		if ("deleteOneCou".equals(action)) { // 來自listAllEmp.jsp
 
 			List<String> errorMsgs = new LinkedList<String>();
 			// Store this set in the request scope, in case we need to
@@ -349,22 +277,44 @@ public class phyCouPromotionServlet extends HttpServlet {
 			req.setAttribute("errorMsgs", errorMsgs);
 	
 				/***************************1.接收請求參數***************************************/
+				Integer course_no = Integer.valueOf(req.getParameter("course_no"));
 				Integer project_no = Integer.valueOf(req.getParameter("project_no"));
 				
+				
 				/***************************2.開始刪除資料***************************************/
-//				PhyCouPromotionService phyCouPromotionSvc = new PhyCouPromotionService();
-//				phyCouPromotionSvc.deletePro(project_no);
 				PhyCouPromotionDetailService phyCouPromotionDetailSvc = new PhyCouPromotionDetailService();
-				phyCouPromotionDetailSvc.deleteOnePro(project_no);
+				phyCouPromotionDetailSvc.deleteOneCou(project_no, course_no);
 				
 				/***************************3.刪除完成,準備轉交(Send the Success view)***********/								
-				String url = "/phyCouPromotion/listAllPro.jsp";
+				String url = "/phyCouPromotion/listAllProDetail.jsp";
 				RequestDispatcher successView = req.getRequestDispatcher(url);
 				successView.forward(req, res);
 		}
 	
 	
 //	if ("listPros_ByCompositeQuery".equals(action)) { 
+		if ("deleteOnePro".equals(action)) { // 來自listAllEmp.jsp
+			
+			List<String> errorMsgs = new LinkedList<String>();
+			// Store this set in the request scope, in case we need to
+			// send the ErrorPage view.
+			req.setAttribute("errorMsgs", errorMsgs);
+			
+			/***************************1.接收請求參數***************************************/
+			Integer project_no = Integer.valueOf(req.getParameter("project_no"));
+			
+			
+			/***************************2.開始刪除資料***************************************/
+			PhyCouPromotionDetailService phyCouPromotionDetailSvc = new PhyCouPromotionDetailService();
+			phyCouPromotionDetailSvc.deleteOnePro(project_no);
+			
+			/***************************3.刪除完成,準備轉交(Send the Success view)***********/
+//			String url = req.getContextPath()+"/phyCouPromotion/listAllProDetail.jsp";
+			String url = "/phyCouPromotion/listAllProDetail.jsp";
+			RequestDispatcher successView = req.getRequestDispatcher(url);
+			successView.forward(req, res);
+		}
+		
 //		List<String> errorMsgs = new LinkedList<String>();
 //		// Store this set in the request scope, in case we need to
 //		// send the ErrorPage view.
